Move `activateAiStepReducer` to own namespace to keep code modre clean and readable
